# Fast implementation of the minimum fill ordering heuristic.