--- a/Project.toml
+++ b/Project.toml
@@ -21,7 +21,7 @@ StarAlgebras = "0c0c59c1-dc5f-42e9-9a8b-b5dc384a6cd1"
 SymbolicWedderburn = "858aa9a9-4c7c-4c62-b466-2421203962a2"
 
 [compat]
-CliqueTrees = "0.4"
+CliqueTrees = "0.5"
 DataStructures = "0.18"
 JuMP = "1.10"
 MathOptInterface = "1.13"