I have a library [CliqueTrees.jl](https://github.com/AlgebraicJulia/CliqueTrees.jl/tree/main) with a fast implementation of the minimum fill ordering heuristic. Here are some benchmarks.

```julia
using BenchmarkTools
using MatrixMarket
using SuiteSparseMatrixCollection
using SparseArrays
using SumOfSquares.Certificate.Sparsity: ChordalExtensionGraph as CEG

ssmc = ssmc_db()
name = "bcsstk13"
matrix = mmread(joinpath(fetch_ssmc(ssmc[ssmc.name .== name, :], format="MM")[1], "$(name).mtx"));

G = CEG.LabelledGraph{Int}()

for j in axes(matrix, 2)
    CEG.add_node!(G, j)
    
    for i in @view rowvals(matrix)[nzrange(matrix, j)]
        if i < j
            CEG.add_edge!(G, i, j)
        end
    end
end
```

**current implementation**

```julia
julia> @benchmark CEG.completion(G, CEG.ChordalCompletion())
BenchmarkTools.Trial: 1 sample with 1 evaluation per sample.
 Single result which took 12.870 s (1.42% GC) to evaluate,
 with a memory estimate of 648.32 MiB, over 892676 allocations.
```

The completed graph has 411703 edges.

```julia
julia> H, _ = CEG.completion(G, CEG.ChordalCompletion()); sum(length, H.graph.neighbors)
823406
```

**new implementation**

```julia
@benchmark CEG.completion(G, CEG.CliqueTreesCompletion())
BenchmarkTools.Trial: 36 samples with 1 evaluation per sample.
 Range (min … max):  133.553 ms … 277.619 ms  ┊ GC (min … max): 0.00% … 51.80%
 Time  (median):     135.738 ms               ┊ GC (median):    1.35%
 Time  (mean ± σ):   140.293 ms ±  23.719 ms  ┊ GC (mean ± σ):  4.25% ±  8.63%

  █▄                                                             
  ██▁▃▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃ ▁
  134 ms           Histogram: frequency by time          278 ms <

 Memory estimate: 35.07 MiB, allocs estimate: 49814.
 ```
 
 The completed graph has 249445 edges.
 
 ```julia
julia> H, _ = CEG.completion(G, CEG.CliqueTreesCompletion()); sum(length, H.graph.neighbors)
498890
```

You can take a look at the implementation [here](https://github.com/AlgebraicJulia/CliqueTrees.jl/blob/b7b64bcc3afb60b5aadfa0cc1ba8b0b145b3443b/src/elimination_algorithms.jl#L1065).