@samuelsonric 2025-02-23
Here, I am benchmarking two examples from section 7.1 of [this paper](https://arxiv.org/abs/1912.08899).

```julia
using BenchmarkTools, DynamicPolynomials, Random, SumOfSquares
using SumOfSquares.Certificate.Sparsity: ChordalExtensionGraph as CEG
using CSDP: Optimizer

solver = Optimizer
Random.seed!(100)

# Network Problem 1
function network_1(n)
    a = rand(n) .+ 1.0
    b = (rand(n, n) .+ 0.5) ./ n
    
    @polyvar x[1:n]
    
    return sum(1:n) do i
        a[i] * (x[i]^2 + x[i]^4) - sum(1:n) do k
            b[i, k] * x[i]^2 * x[k]^2
        end
    end
end

# Broyden Banded Function
function broyden_banded(n)
    @polyvar x[1:n]
    
    return sum(1:n) do i
        f = x[i] * (2 + 5x[i]^2) + 1
        
        for j in max(1, i - 5):min(n, i + 1)
            if i != j
                f += (1 + x[j]) * x[j]
            end
        end
        
        return f^2
    end
end

function sos_min(poly, sparsity)
    model = Model(solver)
    set_attribute(model, "printlevel", 0)
    @variable(model, t)
    @objective(model, Max, t)
    con_ref = @constraint(model, poly - t in SOSCone(), sparsity = sparsity)
    optimize!(model)
    return value(t)
end
```

## Network Problem 1

```julia
julia> f = network_1(60);
```

**current implementation**

```julia
julia> @time sos_min(f, Sparsity.Monomial(CEG.ChordalCompletion()))
13.011651 seconds (115.73 M allocations: 17.944 GiB, 35.21% gc time)
-1.24354693209483e-12
```

**new implementation**

```julia
julia> @time sos_min(f, Sparsity.Monomial(CEG.CliqueTreesCompletion()))
12.983566 seconds (115.72 M allocations: 17.945 GiB, 37.04% gc time)
-1.2436857099729082e-12
```

No difference.

## Broyden Banded Function

```julia
julia> f = broyden_banded(10);
```

**current implementation**

```julia
julia> @time sos_min(f, Sparsity.Monomial(CEG.ChordalCompletion()))
105.182118 seconds (25.52 M allocations: 1.446 GiB, 0.37% gc time, 0.02% compilation time)
-3.8924696799114145e-11
```

**new implementation**

```julia
julia> @time sos_min(f, Sparsity.Monomial(CEG.CliqueTreesCompletion()))
7.179278 seconds (7.27 M allocations: 456.838 MiB, 0.57% gc time)
-9.76190517310016e-11
```

Substantially improved!