@blegat 2025-02-24
Thanks a lot for the PR with detailed benchmarks. The speedup is quite impressive.
I don't think it makes sense to maintain our separate implementation of these algorithms now.
Can you remove the current `ChordalCompletion` and rename `CliqueTreesCompletion` as `ChordalCompletion` ? You can also remove `AbstractGreedyAlgorithm` and the whole implementation.

Just to clarify, do you know why you are computing a better solution ? Is this because you use a different heuristic ? Or our implementation is buggy ?

About the failure on Julia v1.6, don't worry, you can just rebase once https://github.com/jump-dev/SumOfSquares.jl/pull/395 is merged