@samuelsonric 2025-02-28
@blegat I you don't mind I would like some help dealing with the documentation and examples.