@blegat 2025-03-06
Nice, all the dependencies are also quite standard packages